+++ test/src/terminal_test.dart
@@ -49,7 +49,7 @@ void main() {
         CellOffset(10, 10),
       );
 
-      expect(output, ['\x1B[M +,']);
+      expect(output, ['\x1B[M ++']);
     });
   });